added Thoughts.jsx to clean up App.jsx
